ted_gemini/requirements.txt
google-adk
certifi

